Add detailed search story.
